Add standard Ruby ecosystem artifacts to allow developers to install all required dependencies by simply running `bundle install`. Additionally, by declaring these dependencies (and their versions) via a standard Gemfile, we ensure that all developers are using the same version of the dependencies.
For non-Ruby developers, this will make installing the dependencies of
this project easier since setup is now just a single command.

For Ruby developers, this project now plays nicely with their preferred
Ruby environment manager (whether that be RVM or rbenv), ensuring that all
dependencies are installed in an appropriate gemset.