clarify about scopes (#154)
* clarify about scopes

the external reference was misleading as it didn't define any valid/relevant scopes. This change formally defines the system scopes, and clarifies the documentation around that. A later pull request will deal with communication request (it's wrong to mandate that all uses of the backend services involve systems that have clinical users)

* remove the text about required scopes